updated ux based on feedback
