+++ components/auction/auction-detail-ui/AllPayDetail.tsx
@@ -116,6 +116,11 @@ export function AllPayDetail(
                   : "ended"
               }
             />
+            {Date.now() >= Number(currentAuction.deadline) * 1000 && (
+                <p className="text-sm font-medium text-muted-foreground">
+                    {currentAuction.isClaimed ? "Asset has been claimed" : "Asset has not been claimed yet"}
+                </p>
+            )}
           </div>
         </div>
