+++ components/auction/auction-detail-ui/EnglishDetail.tsx
@@ -123,6 +123,11 @@ export function EnglishDetail({
                   : "ended"
               }
             />
+            {Date.now() >= Number(currentAuction.deadline) * 1000 && (
+                <p className="text-sm font-medium text-muted-foreground">
+                    {currentAuction.isClaimed ? "Asset has been claimed" : "Asset has not been claimed yet"}
+                </p>
+            )}
           </div>
         </div>
