+++ components/auction/auction-detail-ui/ExponentialDetail.tsx
@@ -52,6 +52,11 @@ export function ExponentialDetail({
                   : "ended"
               }
             />
+            {Date.now() >= Number(currentAuction.deadline) * 1000 && (
+                <p className="text-sm font-medium text-muted-foreground">
+                    {currentAuction.isClaimed ? "Asset has been claimed" : "Asset has not been claimed yet"}
+                </p>
+            )}
           </div>
         </div>
