+++ components/auction/auction-detail-ui/LogarithmicDetail.tsx
@@ -51,6 +51,11 @@ export function LogarithmicDetail({
                   : "ended"
               }
             />
+            {Date.now() >= Number(currentAuction.deadline) * 1000 && (
+                <p className="text-sm font-medium text-muted-foreground">
+                    {currentAuction.isClaimed ? "Asset has been claimed" : "Asset has not been claimed yet"}
+                </p>
+            )}
           </div>
         </div>
