+++ components/auction/auction-detail-ui/VickreyDetail.tsx
@@ -39,7 +39,9 @@ export function VickreyDetail({
       const auctionService = getAuctionService(currentAuction.protocol);
       const currentBlock = await publicClient.getBlockNumber();
       const fromBlock =
-        currentBlock > BigInt(10000000) ? currentBlock - BigInt(10000000) : BigInt(0);
+        currentBlock > BigInt(10000000)
+          ? currentBlock - BigInt(10000000)
+          : BigInt(0);
       const bidHistory = await auctionService.getBidHistory(
         publicClient,
         BigInt(auctionId),
@@ -150,6 +152,13 @@ export function VickreyDetail({
                 getVickreyPhase(currentAuction) === "ended" ? "ended" : "active"
               }
             />
+            {Date.now() >= Number(currentAuction.bidRevealEnd) * 1000 && (
+              <p className="text-sm font-medium text-muted-foreground">
+                {currentAuction.isClaimed
+                  ? "Asset has been claimed"
+                  : "Asset has not been claimed yet"}
+              </p>
+            )}
           </div>
         </div>
