+++ lib/contract-data.ts
@@ -2531,7 +2531,7 @@ export const VICKREY_ABI = [
   ] as const;
 
 export const AUCTION_CONTRACTS = {
-  English: "0xA7eCB5B2dd3cD9b5326812449AEA89F478570c19",
+  English: "0x454094B29146D815dea199a7Ce5A14945AE66a77",
   AllPay: "0x00Ac82D482e3B918C22706D429a25c91324aF45b",
   Exponential: "0x4dE6AbeccaE5F63C8963616e42397ab41ac1840a", 
   Linear: "0xF54f9f0918A33772DC0984F058dd0aeed8fE4230",
